finish script to add admin proposal
